fix: 1280 * 720 위치로 변환 (#261)

+++ app/src/main/java/com/example/mhnfe/ui/screens/monitoring/kvs/AiViewModel.kt
@@ -29,7 +29,7 @@ class AiViewModel @Inject constructor(
 
     fun processFrame(
         bitmap: Bitmap?,
-        onResult: (Int, String, Float, List<Map<String, Float>>) -> Unit
+        onResult: (Int, String, Float, List<Map<String, Int>>) -> Unit
     ) {
         viewModelScope.launch {
             try {
